Checkpoint - Basic structure of DataProviders & Dataset objects
