+++ tests/test_examples_jupyter.py
@@ -17,8 +17,7 @@ import numpy as np
 
 from nbconvert.exporters import PythonExporter
 
-from gordo_components.dataset.datasets import DataLakeBackedDataset
-
+from gordo_components.data_provider.providers import DataLakeBackedDataset
 
 logger = logging.getLogger(__name__)
